requirements.txt
flask~=3.1.0

waitress~=3.0.2
pandas~=2.2.3
numpy~=1.26.4
pillow~=11.1.0
matplotlib~=3.10.1
pygame~=2.6.1